Change SQL examples
SQL backend only allows SELECT right now. Removing SELECT from the
query input box is the simplest fix.

An alternative is allowing SELECT still, but just removing this
from the string that is sent to the backend. This has greater
possibility for not covering all the possible cases, so rejected this
idea for now.